Refactor for Shared Dependencies
