+++ src/api/deps.py
@@ -1,0 +1,9 @@
+from src.data.database import SessionLocal
+
+
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
