+++ src/api/routers/health.py
@@ -1,0 +1,16 @@
+from fastapi import APIRouter, Depends
+from sqlalchemy import text
+from sqlalchemy.orm import Session
+
+from src.api.deps import get_db
+
+router = APIRouter()
+
+
+@router.get("/health", tags=["Health"])
+def health_check(db: Session = Depends(get_db)):
+    try:
+        db.execute(text("SELECT 1"))
+        return {"status": "ok", "database_status": "ok"}
+    except Exception as e:
+        return {"status": "ok", "database_status": "error", "detail": str(e)}
